計算ロジックを外部に移動
