+++ src/components/utils/base/mcModel/scripts/main.ts
@@ -198,6 +198,35 @@ function modelFaceToModelProp(face: {
   };
 }
 
+export type ModelFaces = {
+  texture: string;
+  matrix3d: [
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number,
+    number
+  ];
+  brightness: {
+    // brightness = x => base + amp * sin(x + phase * 2 * pi)
+    base: number; // 明るさの基礎値 [%]
+    amp: number; // 明るさの振幅 (右に来た時と左に来た時で明るさがどのくらい変わるか) [%]
+    phase: number; // 明るさの位相 (アニメーション開始時の位相) [0..1]
+  };
+  xywh: [number, number, number, number];
+}[];
+
 export function resolveModelFaces(model: McModel) {
   // TODO: ここでparentとtexture変数を解決
